pyproject.toml
[tool.poetry]
name = "mppsolar"
version = "0.16.30"
description = "Package to communicate with Solar inverters and BMSs"
authors = ["John Blance"]
readme = "README.md"
packages = [
    {include = "mppsolar"},
    {include = "powermon"},
    {include = "api"},
    {include = "powermon_api"},
]

[tool.poetry.dependencies]
python = ">=3.10,<4.00"
paho-mqtt = ">=1.6,<2.0.0"
pyserial = ">3"
pydantic = ">=1.8.0"

# Extra dependencies
bleak = { version = "^0.20.2", optional = true }
bluepy = { version = "^1.3.0", optional = true }
cysystemd = { version = ">=1.6.0", optional = true }
fastapi = { version = ">=0.68.0", optional = true }
fastapi-mqtt = { version = ">=1.1", optional = true }
# paho-mqtt = { version = ">=1.6", optional = true }
pymongo = { version = ">=4.6.1", optional = true }
psycopg2-binary = { version = ">=2.9.9", optional = true }
PyYAML = { version = ">=6", optional = true }
requests = { version = ">=2.31.0", optional = true }
sqlalchemy = { version = "^2.0.15", optional = true }
strenum = { version = "^0.4.1", optional = true }
uvicorn = { version = ">=0.15.0", optional = true }
twine = { version = "^4.0.2", optional = true }
#flake8 = { version = "^6.0.0", optional = true }
#black = { version = "^23.3.0", optional = true }
#coverage = { version = "^7.2.5", optional = true }
construct = { version = "^2.10.70", optional = true }
pymodbus = { version = "^3.6.4", optional = true }
pycryptodome = { version = "^3.20.0", optional = true }
pyaml-env = { version = "^1.2.1", optional = true }



[tool.poetry.extras]
api = ["fastapi", "fastapi-mqtt", "sqlalchemy", "uvicorn"]
ble = ["bleak", "bluepy", "pycryptodome"]
dev = ["twine", "bleak", "bluepy", "PyYAML", "strenum", "construct", "pymodbus", "pycryptodome", "pyaml-env"]
# mqtt = ["paho-mqtt"]
mongo = ["pymongo"]
pgsql = ["psycopg2-binary"]
powermon = ["PyYAML", "strenum", "construct", "pyaml-env"]
powermon-ble = ["bleak", "bluepy", "PyYAML", "strenum", "construct", "pyaml-env"]
push = ["requests"]
systemd = ["cysystemd"]
modbus = ["pymodbus"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mpp-solar = "mppsolar:main"
mppsolar = "mppsolar:main"
jkbms = "mppsolar:main"
powermon = "powermon:main"
powermon-api = "powermon_api.api:main"

[tool.black]
line-length = 149
